Run both JUnit 4 and 5 tests
Current DetectorPlugin only run JUnit 4 tests because there is
only one Runner for JUnit 4. With the update of testrunner plugin,
it can create another Runner for JUnit 5.

So try to run both Runners.

There are some issues with the output folders for the runner.
Original there is only one Runner, so one type of output folders
or output files is enough. But since we may run two Runners,
different results of those Runners need to be saved into seperate
places. There are still some issues, which will be fixed in the
next patches.